chore: extract function to utils
